Remove episodes which will not be used
Last time I taught the python class, I was not able to touch too much
the three last episodes (Defensive Programing, Debugging and Command
Line).

In addition to that, I don't think that the defensive part is
appropriate (assert in software is not a good things for my point of
view, it is in test development not in production where the

try:
except ... :
finally:

is better.

The debugging does not show how to do it, just describe some method how
to do it.

The command line is outdated. Nobody will use sys.argv nowadays.
Developpers would prefer somethink like Click or docopt.